Top-up: closing_balance is a decimal, and M35 is topup_extra
Two model corrections the redesign needs before its screens can be right.

closing_balance parsed as int truncated 86217.08 to 86217 — the same defect
yield/collection_fee/penalty_fee carried until 2026-09-11, and with the same
two effects: a figure misstated on screen and a changed transfer_amount on the
submit body. The redesigned card and amount screens both show it to two
decimals under a total that has to add up, so this is now load-bearing.
TopupFlow.closingBalance and payoutAmount widen with it.

applySpecialLimit reads topup_extra, not topup_specials (instructed: 'วงเงิน
พิเศษ M35 -> is topup_extra'), and topup_special_flag is no longer the gate — a
non-zero amount is. That flag belonged to topup_specials, so requiring it would
hide money the backend granted. Three tests pin the new field, the dropped
gate, and that topup_specials alone no longer uplifts anything.

+++ lib/p_loan/application/models/loan_contract.dart
@@ -194,7 +194,13 @@ class ContractDetails {
   });
 
   /// Outstanding principal, deducted from the new loan amount at payout.
-  final int closingBalance;
+  /// **`double`, not `int`** — the wire sends `86217.08` and the redesigned
+  /// card and amount screens both put this figure on screen to two decimals,
+  /// under a total that has to add up. Truncating it misstated the deduction
+  /// by up to a baht *and* changed `transfer_amount` on the submit body, the
+  /// same defect `yield`/`collection_fee`/`penalty_fee` carried until
+  /// 2026-09-11. Widened 2026-09-12 for the same reason.
+  final double closingBalance;
   final String comcode;
   final String branchCode;
   final String branchName;
@@ -229,7 +235,7 @@ class ContractDetails {
 
   factory ContractDetails.fromJson(Map<String, dynamic> json) =>
       ContractDetails(
-        closingBalance: asInt(json['closing_balance']),
+        closingBalance: asDouble(json['closing_balance']),
         comcode: asString(json['comcode']),
         branchCode: asString(json['branch_code']),
         branchName: asString(json['branch_name']),
